Standardize on structlog logging
Rather than having cog.director and cog.server.http emit completely
different log formats, extract the logging configuration into a shared
module and have both entrypoints use the same log configuration.
